+++ js/src/api/format/input.js
@@ -15,9 +15,9 @@
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
 import BigNumber from 'bignumber.js';
-import { range } from 'lodash';
 
 import { isArray, isHex, isInstanceOf, isString } from '../util/types';
+import { padLeft } from '../util/format';
 
 export function inAddress (address) {
   // TODO: address validation if we have upper-lower addresses
@@ -51,16 +51,6 @@ export function inHash (hash) {
   return inHex(hash);
 }
 
-export function padRight (input, length) {
-  const value = inHex(input).substr(2, length * 2);
-  return '0x' + value + range(length * 2 - value.length).map(() => '0').join('');
-}
-
-export function padLeft (input, length) {
-  const value = inHex(input).substr(2, length * 2);
-  return '0x' + range(length * 2 - value.length).map(() => '0').join('') + value;
-}
-
 export function inTopics (_topics) {
   let topics = (_topics || [])
     .filter((topic) => topic === null || topic)
@@ -75,10 +65,6 @@ export function inTopics (_topics) {
 
       return padLeft(topic, 32);
     });
-
-  // while (topics.length < 4) {
-  //   topics.push(null);
-  // }
 
   return topics;
 }
